Execepciones

+++ database/migrations/2024_03_17_022553_create_productos_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -15,7 +14,7 @@ return new class extends Migration
             $table->id();
             $table->string('codigo', 100)->unique();
             $table->string('descripcion', 100);
-            $table->double('precioSugerido', 12,2);
+            $table->decimal('precioSugerido', 12, 2);
             $table->integer('idCategoria')->nullable();
             $table->boolean('aplicaStock');
             $table->timestamps();
