+++ database/migrations/2024_03_17_051815_create_detalles_documentos_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -15,7 +14,7 @@ return new class extends Migration
             $table->id();
             $table->integer('idDocumento');
             $table->integer('idProducto');
-            $table->double('cantidad', 12,2);
+            $table->decimal('cantidad', 12, 2);
             $table->integer('idUsuario');
             $table->timestamps();
         });
