+++ database/migrations/2024_03_17_054001_create_movimiento_jornadas_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -15,7 +14,7 @@ return new class extends Migration
             $table->id();
             $table->integer('idJornada');
             $table->integer('tipo');
-            $table->double('dinero', 12,2);
+            $table->decimal('dinero', 12, 2);
             $table->integer('idUsuario');
             $table->string('descripcion', 255)->nullable();
             $table->timestamps();
